Item 6

+++ src/com/company/Auto.java
@@ -1,7 +1,7 @@
 package com.company;
 
 public class Auto extends Vehiculo {
-    public Auto(Double velocidad, Double aceleracion, Double anguloDeGiro, String patente) {
+    public Auto(Integer velocidad, Integer aceleracion, Integer anguloDeGiro, String patente) {
         super(velocidad, aceleracion, anguloDeGiro, patente, 1000, 4);
     }
 }
