+++ src/com/company/Moto.java
@@ -2,7 +2,7 @@ package com.company;
 
 public class Moto extends Vehiculo {
 
-    public Moto(Double velocidad, Double aceleracion, Double anguloDeGiro, String patente) {
+    public Moto(Integer velocidad, Integer aceleracion, Integer anguloDeGiro, String patente) {
         super(velocidad, aceleracion, anguloDeGiro, patente, 300, 2);
     }
 }
